+++ foo_vis_milk2/resource.h
@@ -935,6 +935,9 @@
 #define IDC_FORMAT_INFO_USE_ARTWORK     1133
 #define IDC_FORMAT_INFO_EXAMPLE_2_TEXT  1134
 #define IDC_CB_SEPARATE_CLICK_PROMPT_FONT 1135
+#define IDC_OPEN_MILKDROP_FOLDER        1136
+#define IDC_OPEN_TEXTURES_FOLDER        1137
+#define IDC_PREFS_SCROLLBAR             1138
 #define ID_SHOWHELP                     40007
 #define ID_SHOWPLAYLIST                 40008
 #define ID_QUIT                         40009
@@ -947,7 +950,7 @@
 #define _APS_NO_MFC                     1
 #define _APS_NEXT_RESOURCE_VALUE        105
 #define _APS_NEXT_COMMAND_VALUE         40011
-#define _APS_NEXT_CONTROL_VALUE         1136
+#define _APS_NEXT_CONTROL_VALUE         1139
 #define _APS_NEXT_SYMED_VALUE           101
 #endif
 #endif